Defer to go-update for the heavy lifting
This is a big refactor of implementation, and only exposes new external APIs. The go-update.v0 library handles a bsdiff
patch, so this just checks for an error there and tries a full binary download and update (also via go-update). There's
probably still room for improvement here, but it cleans up a lot of unnecessary code.

This change also adds publicly available Update and UpdateAvailable methods to make this more useful as a library in
need of controlled changes.